PARTE 1/flask_gemini_minimo/requirements.txt
Flask==3.0.3
google-genai==1.35.0
python-dotenv==1.0.1
gunicorn==23.0.0
google-cloud-storage==3.1.1
pypdf==5.4.0
numpy==2.2.4
